Fix App.tsx logic related to authentication
### Summary

Before this fix, the log out button doesn't quite work. This diff fixes this problem and makes the login and permission check login in App.tsx much cleaner.

Instead of mixing the code for isLogged in status check and professor view permission check, the new version keeps the concerns separate.

The old code has the bad behavior that an unlogged in user will be stuck in the loading state forever if he's trying to access professor view. This diff now corrects the behavior by directing the unauthenticated user to the login screen.

### Test Plan

Try click logout. Things should work automatically.