Filter sessions by date
On the client side, we still fetch all the sessions, but we filter it by date to feed into CalendarSessions component.
Since we are only fetching data for one course, and we don't pull the entire giant list, there won't be big performance hit.
On the bright side, it makes implementation significantly easier!