Store professor and ta list in `FireCourse`
Firebase security rules do not let you do query; instead, you can only get document by id or check exists. Our current setup requires to perform a query in `courseUsers` collection to check whether current auth user is a professor of a certain class, which is unachievable with firebase security rule.
The only solution is to duplicate the data in `courses` collection. Since number of professors and tas should be small, it should not create a huge performance overhead. It even introduces a benefit: now with the duplicated data, the query in `ProfessorRolesTable` can be implemented more efficiently.