More efficient firebase hooks by reusing existing data
A few diffs ago, I setup the singleton observable infra. The idea is that the query is only performed once, and make other resources that depend on this data subscribe to a single source of truth.
This diff finishes up the performance optimization inside `firehook.ts`. It caches all the fetches courses, and only do filter based on already fetched data.