Completely implement student enrollment model
### Summary

This diff completely implemnents the backend support for student self-enroll model.
Frontend will automatically compute a list of classes to enroll and unenroll based on user's selected list and a list of all classes. It can correctly handle the case that professor/TAs cannot unenroll themselves from the interface.

### Notes

With the firebase reimplementation, saving can happen instantaneously. Do we still need a save and cancel button?