Merge users and courseUsers collection (#250)
### Rationale

To improve performance of the questions container, this diff is a must: we should be able to prefetch a list of all course users along with user information in one request. Thus, `users` and `courseUsers` collection must be merged.

### Implementation

The implementation is constrained by several facts:

- Enforcibility of security rules
- Efficiency of query

Due to some weird constraint of firebase security rules, it's not possible to store the role and course enrollment together. Therefore, I have to split it to `courses` and `roles` field. If we store the `student` field, it's impossible to enforce the rule that only professsor can edit role. Therefore, I made the role == `student` implicit. I documented all the invaraints in `fireData.d.ts`.