Let TA add comments to student question (#267)
During office hour today, Scott mentions that we might want a virtual whiteboard.

Implementing one is hard, but provide link to one is easy: we add an optional field to question called `taComment`.TA can edit this field, and it will be presented to the students. We can implement notification for this in this way in the future:

```typescript
useEffect(() => {
  if (cachedOldComment !== newComment && isStudent) {
    fireNotification(...);
  }
}, [cachedOldComment, newComment, isStudent]);
```

Right now the GUI does not look good, so it's gates behind a flag. You must enable it by manually running `localStorage.setItem('FEATURE_TA_COMMENT_ENABLE_FLAG', 'true')` in your console.

This should not be merged until we got PM and designer approval.