Fix the problem of infinite redirect (#283)
### Summary

When the only course you enrolled in is expired, it will try to redirect to default course. If you are not in that course or the course is invalid, it will redirect to the same page again, causing an infinite redirect that breaks the site.

This diff make it redirect to the edit page instead.

### Test Plan

1. Make yourself ONLY enrolled in the expired INFO4998.
2. Open the page, it will redirect you to edit.

### Notes

After this diff is merged, let's change INFO4998's end date to June.