Bootstrap admin view
### Summary

This is a preparatory diff to implement a full functional admin view. The eventual goal is to implement an admin view that can creating an editing courses through UI.

This diff implements the reads of course data, and relevant professor and TA data. These data are already publicly available, so I didn't add any new security checks. The UI is ugly, but it's only for internal use so it doesn't matter.

The editing functionality is prepared for the next diff.

### Test Plan

Visit `/admin`, you can see all courses and TA professor information.