Support creating and editing some fields of an existing course in admin page
### Summary

This diff allows user to edit certain fields of an existing class. Fields that cannot be edited are the fields that can be edited on the professor side, or start and end date.

I also updated the firebase security rule that allows an admin user to edit course. Right now, I hardcoded the emails of Scott and I in that rule.

### Test Plan

Try to edit the dummy CS 3110 class or create a new class. The UX is not good but it works.