AdminView (#288)
* Bootstrap admin view

### Summary

This is a preparatory diff to implement a full functional admin view. The eventual goal is to implement an admin view that can creating an editing courses through UI.

This diff implements the reads of course data, and relevant professor and TA data. These data are already publicly available, so I didn't add any new security checks. The UI is ugly, but it's only for internal use so it doesn't matter.

The editing functionality is prepared for the next diff.

### Test Plan

Visit `/admin`, you can see all courses and TA professor information.

* Support creating and editing some fields of an existing course in admin page

### Summary

This diff allows user to edit certain fields of an existing class. Fields that cannot be edited are the fields that can be edited on the professor side, or start and end date.

I also updated the firebase security rule that allows an admin user to edit course. Right now, I hardcoded the emails of Scott and I in that rule.

### Test Plan

Try to edit the dummy CS 3110 class or create a new class. The UX is not good but it works.