Fix linter errors
### Summary

Fix some easy linter errors.

Some of the a11y fixes are tricky, since I have to restyle something. It appears to me that the original authors are just abusing the default style provided by span, div, without considering that we need to use `button` to convey the semantic meaning of button to make the website accessible.

### Test Plan

Less linter errors in CI. Changed styles are identical to the old ones.