Add "virtual location" for TAs
This adds a "session profile" document for each TA which stores information about them per-session. Right now this is used to store their current virtual location (likely a Zoom link). When they are set as an answerer this value is copied into the question.

TODO: Allow TAs to update locations once they are the answerer
TODO: Show the virtual location in the student view