Add static script to force migration.
Our old instance of QueueMe.in didn't set its cache headers correctly so for some people index.html is permanently cached. To fix this, we're going to insert a script which will forcibly remove the service worker and reload the instance.

The script is harmless, unless you have an old install.